pyproject.toml
[project]
name = "mindcare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.1.1",
]
